+++ gcpy/benchmark/run_benchmark.py
@@ -70,8 +70,7 @@ from gcpy.benchmark.modules.benchmark_funcs import \
     make_benchmark_mass_accumulation_tables, \
     make_benchmark_operations_budget, \
     make_benchmark_collection_2d_var_plots, \
-    make_benchmark_collection_3d_var_plots, \
-    make_benchmark_operations_budget
+    make_benchmark_collection_3d_var_plots
 from gcpy.benchmark.modules.ste_flux import make_benchmark_ste_table
 from gcpy.benchmark.modules.oh_metrics import make_benchmark_oh_metrics
 from gcpy.benchmark.modules.run_1yr_fullchem_benchmark \
@@ -488,8 +487,8 @@ def run_benchmark_default(config):
                 gcc_vs_gcc_refstr,
                 dev,
                 gcc_vs_gcc_devstr,
+                colname,
                 spcdb_files,
-                colname=colname,
                 var_prefix=colname,
                 dst=gcc_vs_gcc_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -515,8 +514,8 @@ def run_benchmark_default(config):
                 gcc_vs_gcc_refstr,
                 dev,
                 gcc_vs_gcc_devstr,
+                colname,
                 spcdb_files,
-                colname=colname,
                 dst=gcc_vs_gcc_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
                 overwrite=True,
@@ -530,13 +529,14 @@ def run_benchmark_default(config):
             print("\n%%% Creating GCC vs. GCC 2D StateMet plots %%%")
 
             # Create plots
+            colname = "StateMet"
             make_benchmark_collection_2d_var_plots(
                 refmet,
                 gcc_vs_gcc_refstr,
                 devmet,
                 gcc_vs_gcc_devstr,
+                colname,
                 spcdb_files,
-                colname="StateMet",
                 var_prefix='Met',
                 dst=gcc_vs_gcc_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -551,13 +551,14 @@ def run_benchmark_default(config):
             print("\n%%% Creating GCC vs. GCC 3D StateMet plots %%%")
 
             # Create plots
+            colname = "StateMet"
             make_benchmark_collection_3d_var_plots(
                 refmet,
                 gcc_vs_gcc_refstr,
                 devmet,
                 gcc_vs_gcc_devstr,
+                colname,
                 spcdb_files,
-                colname="StateMet",
                 var_prefix='Met',
                 dst=gcc_vs_gcc_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -741,6 +742,7 @@ def run_benchmark_default(config):
                 dev,
                 gcc_ref_sec_diff,
                 gcc_dev_sec_diff,
+                spcdb_files,
                 benchmark_type=config["options"]["bmk_type"],
                 label=comparison_str,
                 dst=gcc_vs_gcc_tablesdir,
@@ -876,7 +878,7 @@ def run_benchmark_default(config):
                 overwrite=True,
                 verbose=False,
             )
-            
+
 
     # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
     # Create GCHP vs GCC benchmark plots and tables
@@ -1082,8 +1084,8 @@ def run_benchmark_default(config):
                 gchp_vs_gcc_refstr,
                 dev,
                 gchp_vs_gcc_devstr,
+                colname,
                 spcdb_files,
-                colname=colname,
                 var_prefix=colname,
                 dst=gchp_vs_gcc_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -1114,8 +1116,8 @@ def run_benchmark_default(config):
                 gchp_vs_gcc_refstr,
                 dev,
                 gchp_vs_gcc_devstr,
+                colname,
                 spcdb_files,
-                colname=colname,
                 dst=gchp_vs_gcc_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
                 overwrite=True,
@@ -1129,13 +1131,14 @@ def run_benchmark_default(config):
             print("\n%%% Creating GCHP vs. GCC 2D StateMet plots %%%")
 
             # Create plots
+            colname = "StateMet"
             make_benchmark_collection_2d_var_plots(
                 refmet,
                 gchp_vs_gcc_refstr,
                 devmet,
                 gchp_vs_gcc_devstr,
+                colname,
                 spcdb_files,
-                colname="StateMet",
                 var_prefix='Met',
                 dst=gchp_vs_gcc_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -1150,13 +1153,14 @@ def run_benchmark_default(config):
             print("\n%%% Creating GCHP vs. GCC 3D StateMet plots %%%")
 
             # Create plots
+            colname = "StateMet"
             make_benchmark_collection_3d_var_plots(
                 refmet,
                 gchp_vs_gcc_refstr,
                 devmet,
                 gchp_vs_gcc_devstr,
+                colname,
                 spcdb_files,
-                colname="StateMet",
                 var_prefix='Met',
                 dst=gchp_vs_gcc_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -1397,6 +1401,7 @@ def run_benchmark_default(config):
                 dev,
                 gcc_dev_sec_diff,
                 gchp_dev_sec_diff,
+                spcdb_files,
                 benchmark_type=config["options"]["bmk_type"],
                 label=comparison_str,
                 operations=[
@@ -1474,7 +1479,7 @@ def run_benchmark_default(config):
                 verbose=False,
                 dev_gchp=True
             )
-            
+
     # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
     # Create GCHP vs GCHP benchmark plots and tables
     # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -1708,8 +1713,8 @@ def run_benchmark_default(config):
                 gchp_vs_gchp_refstr,
                 dev,
                 gchp_vs_gchp_devstr,
+                colname,
                 spcdb_files,
-                colname=colname,
                 var_prefix=colname,
                 dst=gchp_vs_gchp_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -1745,8 +1750,8 @@ def run_benchmark_default(config):
                 gchp_vs_gchp_refstr,
                 dev,
                 gchp_vs_gchp_devstr,
+                colname,
                 spcdb_files,
-                colname=colname,
                 dst=gchp_vs_gchp_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
                 overwrite=True,
@@ -1760,13 +1765,14 @@ def run_benchmark_default(config):
             print("\n%%% Creating GCHP vs. GCHP 2D StateMet plots %%%")
 
             # Create plots
+            colname = "StateMet"
             make_benchmark_collection_2d_var_plots(
                 refmet,
                 gchp_vs_gchp_refstr,
                 devmet,
                 gchp_vs_gchp_devstr,
+                colname,
                 spcdb_files,
-                colname="StateMet",
                 var_prefix='Met',
                 dst=gchp_vs_gchp_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -1781,13 +1787,14 @@ def run_benchmark_default(config):
             print("\n%%% Creating GCHP vs. GCHP 3D StateMet plots %%%")
 
             # Create plots
+            colname = "StateMet"
             make_benchmark_collection_3d_var_plots(
                 refmet,
                 gchp_vs_gchp_refstr,
                 devmet,
                 gchp_vs_gchp_devstr,
+                colname,
                 spcdb_files,
-                colname="StateMet",
                 var_prefix='Met',
                 dst=gchp_vs_gchp_resultsdir,
                 weightsdir=config["paths"]["weights_dir"],
@@ -1909,7 +1916,7 @@ def run_benchmark_default(config):
                 gchp_res=config["data"]["dev"]["gchp"]["resolution"],
                 gchp_is_pre_14_0=config["data"]["dev"]["gchp"]["is_pre_14.0"]
             )
-            
+
             # Date strings for start of run
             ref_date_str = datetime64_to_str(
                 gchp_ref_date,
@@ -2053,6 +2060,7 @@ def run_benchmark_default(config):
                 dev,
                 gchp_ref_sec_diff,
                 gchp_dev_sec_diff,
+                spcdb_files,
                 benchmark_type=config["options"]["bmk_type"],
                 label=comparison_str,
                 operations=[
@@ -2130,7 +2138,7 @@ def run_benchmark_default(config):
                 dst=gchp_vs_gchp_tablesdir,
                 overwrite=True,
             )
-            
+
         # ==================================================================
         # GCHP vs. GCHP summary table
         # ==================================================================
@@ -2192,7 +2200,7 @@ def run_benchmark_default(config):
                 overwrite=True,
                 verbose=False,
             )
-            
+
     # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
     # Create GCHP vs GCC difference of differences benchmark plots
     # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -2203,15 +2211,15 @@ def run_benchmark_default(config):
 
             # Species database files (just use the gcc ones)
             spcdb_files = get_species_database_files(config, "gcc", "gcc")
-            
+
             # Filepaths
             gcc_ref = get_filepath(
-                gcc_vs_gcc_refdir, 
+                gcc_vs_gcc_refdir,
                 "SpeciesConc", 
                 gcc_ref_date
             )
             gcc_dev = get_filepath(
-                gcc_vs_gcc_devdir, 
+                gcc_vs_gcc_devdir,
                 "SpeciesConc", 
                 gcc_dev_date
             )